+++ CHANGE_LOG.md
@@ -1,4 +1,4 @@
 # Change Log
 
-* [2016-11-20] Add `x-forwarded-for` and `x-real-ip` support
+* [2016-11-20] Add `x-forwarded-for`, `x-forwarded-proto` and `x-real-ip` support
 * [2016-10-08] BAD REQUEST with http://www.dtgco.cn/ because it is IIS bug? See https://support.microsoft.com/en-us/kb/2988444